resolved comments

+++ player.py
@@ -8,7 +8,7 @@ class Player:
         self._piece_val = piece_val
 
     def mini_max(
-        self, board: Board, depth: int, alpha: float, beta: float, flag: int
+        self, board: Board, depth: int, alpha: int, beta: int, flag: int
     ) -> tuple[int, Move]:
         if depth == self._max_depth or len(board.is_terminal()) == 0:
             return (board.evaluate(), board.last_move)
